refactor: move statistics api and hooks
